# Reduce allocations in idsForExpr