###### Required for all non-trivial PRs
- [x] Rebased/mergable
- [x] Tests pass

This calculates a max size of the `SeriesIDs` to avoid a bunch of `appends` that would resize it.

```
benchmark                                         old ns/op     new ns/op     delta
BenchmarkMeasurement_SeriesIDForExp_EQRegex-8     67843610      64845446      -4.42%
BenchmarkMeasurement_SeriesIDForExp_NERegex-8     20727611      21460651      +3.54%

benchmark                                         old allocs     new allocs     delta
BenchmarkMeasurement_SeriesIDForExp_EQRegex-8     35             4              -88.57%
BenchmarkMeasurement_SeriesIDForExp_NERegex-8     4              4              +0.00%

benchmark                                         old bytes     new bytes     delta
BenchmarkMeasurement_SeriesIDForExp_EQRegex-8     4655870       804843        -82.71%
BenchmarkMeasurement_SeriesIDForExp_NERegex-8     2408859       2408855       -0.00%
```

cc @jsternberg @e-dard