@e-dard 2016-09-20
Aside from @jsternberg's comments LGTM 👍